requirements.txt
# 项目核心依赖
discord.py
python-dotenv
Pillow
requests
google-genai
PyYAML
chromadb